tools/requirements.txt
pyserial
pyqtgraph
pyqt6
numpy